# docs/swagger.yaml
basePath: /api
definitions:
...
  customer.MembershipPlanRequestDto:
    properties:
      membership_plan_id:
        type: string
      start_date:
        type: string
      status:
        type: string
    required:
    - start_date
    type: object
...
  staff.RequestDto:
    properties:
      is_active:
        type: boolean
      role_name:
        type: string
    required:
    - is_active
    - role_name
    type: object
  staff.ResponseDto:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_active:
        description: Indicates if the staff is still an active employee
        type: boolean
      role_id:
        type: string
      role_name:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
paths:
  /api/staffs:
    get:
      consumes:
      - application/json
      description: Get a list of staff members
      parameters:
      - description: RoleName HubSpotId to filter staff
        example: '"f47ac10b-58cc-4372-a567-0e02b2c3d479"'
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetMemberships of staff members retrieved successfully
          schema:
            items:
              $ref: '#/definitions/staff.ResponseDto'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get a list of staff members
      tags:
      - staff
  /api/staffs/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a staff member by HubSpotId
      parameters:
      - description: Staff HubSpotId
        example: '"f47ac10b-58cc-4372-a567-0e02b2c3d479"'
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: 'No Content: Staff deleted successfully'
        "400":
          description: 'Bad Request: Invalid HubSpotId'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: Staff not found'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Delete a staff member
      tags:
      - staff
    put:
      consumes:
      - application/json
      description: Update a staff member
      parameters:
      - description: Staff HubSpotId
        example: '"f47ac10b-58cc-4372-a567-0e02b2c3d479"'
        in: path
        name: id
        required: true
        type: string
      - description: Staff details
        in: body
        name: staff
        required: true
        schema:
          $ref: '#/definitions/staff.RequestDto'
      produces:
      - application/json
      responses:
        "204":
          description: 'No Content: Staff updated successfully'
        "400":
          description: 'Bad Request: Invalid input'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: Staff not found'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Update a staff member
      tags:
      - staff
...
  /purchases/memberships:
    post:
      consumes:
      - application/json
      description: Allows a customer to purchase a membership plan by providing the
        plan details.
      parameters:
      - description: Membership purchase details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/customer.MembershipPlanRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: Membership purchased successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad Request: Invalid input'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 'Internal Server Error: Failed to process membership purchase'
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Purchase a membership plan
      tags:
      - purchases
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"